+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and removed features go under `Deprecated` and `Removed` respectively.
 - Basic loader to indicate that a query is executing
 - An error page to display errors in query execution
 
+### Changed
+
+- Run query keybinding to `alt-q e` from `shift-q e`
+
 ## [0.0.1] - 2020-12-29
 
 ### Added